Commonize DumpAnchorTest
Part of https://github.com/krzema12/snakeyaml-engine-kmp/issues/269.
