+++ src/commonTest/resources/anchor/issue481.yaml
